requirements.txt
numpy~=1.26.1
PyOpenGL~=3.1.7

# sudo apt-get install freeglut3-dev